deployments/aws/workers/restoration/app/requirements.txt
boto3==1.21.18
botocore==1.24.18
numpy==1.22.2
opencv-python-headless==4.5.5.62
Pillow==9.0.1
rawpy==0.17.1